redpanda_stream/python/streamlit-result/requirements.txt
streamlit==1.54.0
pandas==2.3.3
pyarrow==23.0.1
python-dotenv==1.2.1
s3fs==2026.2.0
tabulate==0.9.0
minio==7.2.20
streamlit-autorefresh==1.0.1